separate singleton and raft unittest into different files
