+++ framework/core/singleton.py
@@ -1,9 +1,5 @@
 #!/usr/bin/env python3
 
-import argparse
-import sys
-import argparse
-import unittest
 from framework.core.decodeParams import decodeParams
 
 
@@ -43,33 +39,3 @@ class Singleton:
     def config(self, config: dict):
         self._config = config
     
-
-SINGLETON = Singleton(log=None)
-
-class RAFTUnitTestCase(unittest.TestCase):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.deviceManager = SINGLETON.deviceManager
-
-    def main():    
-        parser = argparse.ArgumentParser(description="Run RAFT unit tests.")
-        parser.add_argument('--config', type=str, help='Path to the configuration file')
-        args, remaining_args = parser.parse_known_args()
-
-        if args.config:
-            print(f"Using config file: {args.config}")
-            # Load and set the configuration in the singleton
-            singleton_instance = Singleton(args.config)
-            singleton_instance.config = decodeParams(args.config)
-
-        # Run unit tests 
-        unittest.main(argv=[sys.argv[0]] + remaining_args, exit=False)
-
-
-class RAFTUnitTestMain(unittest.TestCase):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.deviceManager = SINGLETON.deviceManager
-
